+++ src/main/java/net/forcemaster_rpg/item/weapons/KnuckleItem.java
@@ -1,20 +1,14 @@
 package net.forcemaster_rpg.item.weapons;
 
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
-import net.minecraft.item.tooltip.TooltipType;
-import net.minecraft.text.Text;
-import net.minecraft.util.Formatting;
 import net.spell_engine.api.item.weapon.SpellWeaponItem;
-
-import java.util.List;
 
 public class KnuckleItem extends SpellWeaponItem {
     public KnuckleItem(ToolMaterial material, Settings settings) {
         super(material, settings);
     }
 
+    /*
     @Override
     public void appendTooltip(ItemStack stack, TooltipContext context, List<Text> tooltip, TooltipType type) {
         if (Screen.hasShiftDown()) {
@@ -24,4 +18,5 @@ public class KnuckleItem extends SpellWeaponItem {
         }
         super.appendTooltip(stack, context, tooltip, type);
     }
+     */
 }